requirements.txt
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
firebase-admin==7.1.0

python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
tqdm==4.67.1
tenacity==9.1.2

protobuf==5.29.5
grpcio==1.76.0
grpcio-status==1.71.2
